whatsappAutoReplay/requirements.txt
flask
python-dotenv
requests
openai>=1.0.0

